Add giphy data source

+++ features/details/src/main/java/com/sample/features/details/ui/DetailsViewModel.kt
@@ -15,7 +15,7 @@ internal class DetailsViewModel @Inject constructor(
 ) : ViewModel() {
 
     fun getItem(): Flow<ItemUiModel> {
-        val id = savedStateHandle.get<String>("id")!!.toInt()
+        val id = savedStateHandle.get<String>("id")!!
         return repository.getItemBy(id)
     }
 }
